# [FRONTEND] Fix benchmark plotting