# Summary
A [recent PR](https://github.com/openai/triton/pull/2036/files#diff-447aaa7319f4083423e0d1ce5ecbf440e7efe60b5bc527b38f82e982280b98d3) changed the behavior of `triton.testing.Benchmark`, causing the x-axis to be incorrect and the dataframe to contain strings instead of values.
This PR improves the behavior to correctly handle the single `x_name` use case while also supporting the new multiple `x_name` with different values use case. I also added some type annotations and updated the docstring for clarity.

# Testing
To test this I made one vanilla test case that fits the most common single `x_name` pattern, and another from [the use case added in the above-mentioned PR](https://github.com/openai/triton/pull/2036/files#diff-8230ecf751625890f518ea30c064b509c1275be1c3b990a3b621d48aaa854771R149-R174) that required supporting multiple `x_name`s with different values.

## Test code:
```
from testing import *

@perf_report(
    [
        Benchmark(
            x_names=["x"],
            x_vals=list(range(-10, 11)),
            line_arg="positive",
            line_vals=[True, False],
            line_names=["True", "False"],
            styles=[("blue", "-"), ("green", "-")],
            plot_name="foo",
            args={},
        )
    ]
)
def foo(x: float, positive: bool) -> tuple[float, float, float]:
    coeff = 2 if positive else -2
    y = x**2 + coeff * x + 1
    return y, y - 1, y + 1

foo.run(show_plots=True, print_data=True)

@perf_report(
    Benchmark(
        # argument names to use as an x-axis for the plot
        x_names=['M', 'N', 'K'],
        x_vals=[
            [2048, 512, 512],
            [2048, 1024, 1024],
            [2048, 2048, 2048],
            [2048, 4096, 4096],
            [2048, 8192, 8192]
        ],  # different possible values for `x_name`
        line_arg='provider',
        # argument name whose value corresponds to a different line in the plot
        # possible values for `line_arg``
        line_vals=['cublas', 'triton'],
        # label name for the lines
        line_names=["cuBLAS", "Triton"],
        # line styles
        styles=[('green', '-'), ('green', '--'),
                ('blue', '-'), ('blue', '--')],
        ylabel="TFLOPS",  # label name for the y-axis
        plot_name="matmul-performance",
        # name for the plot. Used also as a file name for saving the plot.
        args={},
    )
)
def bar(M, N, K, provider):
    return (M + N + K) * (1 + int(provider == "triton"))

bar.run(show_plots=True, print_data=True)
```

## Previous Behavior
![bad_foo](https://github.com/openai/triton/assets/8582453/f1ba8dbc-e77c-487a-a02a-f4b5f8556e39)
![bad_bar](https://github.com/openai/triton/assets/8582453/2f46e10d-b8b6-4d06-9b75-d20865d89002)

```
foo:
    ['x']  True  False
0   [-10]    81    121
1    [-9]    64    100
2    [-8]    49     81
3    [-7]    36     64
4    [-6]    25     49
5    [-5]    16     36
6    [-4]     9     25
7    [-3]     4     16
8    [-2]     1      9
9    [-1]     0      4
10    [0]     1      1
11    [1]     4      0
12    [2]     9      1
13    [3]    16      4
14    [4]    25      9
15    [5]    36     16
16    [6]    49     25
17    [7]    64     36
18    [8]    81     49
19    [9]   100     64
20   [10]   121     81
matmul-performance:
      ['M', 'N', 'K']  cuBLAS  Triton
0    [2048, 512, 512]    3072    6144
1  [2048, 1024, 1024]    4096    8192
2  [2048, 2048, 2048]    6144   12288
3  [2048, 4096, 4096]   10240   20480
4  [2048, 8192, 8192]   18432   36864
```

## New Behavior
![good_foo](https://github.com/openai/triton/assets/8582453/ac7b028c-933a-4e8a-abaa-e7c3d7c31aef)
![good_bar](https://github.com/openai/triton/assets/8582453/fdf104af-727d-45b2-bdcb-e91ae6d9e9cf)

```
foo:
     x  True  False
0  -10    81    121
1   -9    64    100
2   -8    49     81
3   -7    36     64
4   -6    25     49
5   -5    16     36
6   -4     9     25
7   -3     4     16
8   -2     1      9
9   -1     0      4
10   0     1      1
11   1     4      0
12   2     9      1
13   3    16      4
14   4    25      9
15   5    36     16
16   6    49     25
17   7    64     36
18   8    81     49
19   9   100     64
20  10   121     81
matmul-performance:
        M       N       K   cuBLAS   Triton
0  2048.0   512.0   512.0   3072.0   6144.0
1  2048.0  1024.0  1024.0   4096.0   8192.0
2  2048.0  2048.0  2048.0   6144.0  12288.0
3  2048.0  4096.0  4096.0  10240.0  20480.0
4  2048.0  8192.0  8192.0  18432.0  36864.0
```